fix: reflect comment
change to use serialized binaries
